fix: change foreing keys in migration to BIGINT

+++ migrations/20240926211507-create-password-reset-codes-table.js
@@ -21,7 +21,7 @@ module.exports = {
         defaultValue: Sequelize.NOW // Establece la fecha actual por defecto
       },
       userId: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.BIGINT,
         references: {
           model: 'users',
           key: 'id'
